Translate the executed TestAny and WaitAny requests into Test and Wait ones.
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@9457 48e7efb5-ca39-0410-a469-dd3cf9ba447f

+++ src/simix/smx_network.c
@@ -372,7 +372,7 @@ void SIMIX_pre_comm_test(smx_req_t req)
   }
 }
 
-void SIMIX_pre_comm_testany(smx_req_t req, unsigned int idx)
+void SIMIX_pre_comm_testany(smx_req_t req, int idx)
 {
   unsigned int cursor;
   smx_action_t action;
@@ -380,7 +380,7 @@ void SIMIX_pre_comm_testany(smx_req_t req, unsigned int idx)
   req->comm_testany.result = -1;
 
   if (MC_IS_ENABLED){
-    if((int)idx == -1){
+    if(idx == -1){
       SIMIX_request_answer(req);
     }else{
       action = xbt_dynar_get_as(actions, idx, smx_action_t);
@@ -402,7 +402,7 @@ void SIMIX_pre_comm_testany(smx_req_t req, unsigned int idx)
   SIMIX_request_answer(req);
 }
 
-void SIMIX_pre_comm_waitany(smx_req_t req, unsigned int idx)
+void SIMIX_pre_comm_waitany(smx_req_t req, int idx)
 {
   smx_action_t action;
   unsigned int cursor = 0;
